NOTE: ABI and API break in libunity was required to fix a leak. The affected abi+api was not used any where outside libunity itself.
Revamp the internal re-usage of tiles to use Gee collections instead of GSList which doesn't work super well wth Vala mem management

Discovered a bug in Vala causing it to generate self-refs in object constructors if the object has a delegate member which is set to a lambda or a non-static method on 'this'. This was the root cause of a series of leaks we have been battling for some time. Made the offending places use static methods in stead.