[a11y] Exposing as ACTIVE and FOCUSED the launcher icon selected
Orca doesn't speech out all the selection changes. For instance, it doesn't
speech the selection changes if the role is a tool bar. It worked before
because I used a different role. But the proper role of the launcher
is ATK_ROLE_TOOLBAR. In order to solve that, we expose as FOCUSED the
launcher icon selected. From the pov this is consistent information

Fixes https://bugs.launchpad.net/unity/+bug/727133
Fixes https://bugs.launchpad.net/ubuntu/+source/unity/+bug/729165