Fix app launching in Dash for Wine apps and other apps that are deeply nested in ~/.local/share/applications.
We need to do - to / substitution if we don't find the right file by desktop id. Note that the fix is not 110% tight as certain kinds of folder structures will be ignored (the eager reader should experiment with exotic combinations of nested dirs and -'s in folder names :-)). These exotic folder structure have yet to be observed in the wild though - so hopefully no biggie