Panel service: Add priority flag to indicators
When panel service loads the indicators, it also compute the priority
of the indicator entry itself. That value then can be used by
the panel to correctly order the indicators.

Now panel service orders the indicators following the specifications
defined by the design team, and basically from the end of the screen:
 * indicator-session
 * indicator-datetime
 * indicator-sound
 * network manager (nm-applet)
 * bluetooth manager
 * indicator-messages
 * keyboard layout
 * All the others appindicators
 * The application menus