UnityCore: Support for prioritized indicator entries
When an IndicatorEntry is allocated now, we also consider its priority
flag, as the PanelService sends us that value too.

Also, now we avoid to allocate a new entry at every Sync, but we
try to use the already allocated instance if it exists.
Added also the support for removing an indicator entry when
it's not more synced by the PanelSerive, with signal support.

This avoids that unity will create blank indicator entries, but
actually it only generates real entries to be added to the panel.

Finally, now we use have an indicator entry priority flag, we don't
need to order all the entries per-indicator.