Use and XRegion instead of nux::Geometry to determine the dirty regions that need refreshing.
The problem with using nux::Geometry on a flat bounds is something like this:

-------------------------
|d|                     |
|d|   |bbbbbbbbbb|      |
|     |bbbbbbbbbb|      |
|                     |d|
-------------------------

Will result in 

-------------------------
|d|rrrrrrrrrrrrrrrrrrrrr|
|d|rrr|bbbbbbbbbb|rrrrrr|
|rrrrr|bbbbbbbbbb|rrrrrr|
|rrrrrrrrrrrrrrrrrrrrr|d|
-------------------------

Rather than

-------------------------
|r|                     |
|r|   |bbbbbbbbbb|      |
|     |bbbbbbbbbb|      |
|                     |r|
-------------------------

Causing the blur to be updated when the regions in d are changed. We don't
need to do that.

NOTE: I couldn't find anything in nux that works similarly to X11's regions
so I just used it directly. That can be removed if something else from
nux is suitable.

NOTE2: It seems like the damage regions nux is giving us for our BaseWindows
are HUGE, this needs to be fixed in nux.