Add a workaround to avoid the compiz grid plugin to show the full-screen effect
It happens when you try to restore a maximized window by dragging its
title down. If the mouse drags the window when it's in the panel
compiz draw a fullscreen shape of the window, but that's incorrect
as discussed in bug #838923

A workaround to avoid this is to track the dragging in the panel
and to activate that drag only when we're out from the panel.

So here we simulate a drag, changing the mouse pointer, then when
the mouse drag goes outside the panel, the real window drag is activated.

This doesn't fix completely the bug #838923, since that is mostly caused
by an X race in updating the Window X properties.