PanelService: fix keynavigation on priority based indicators.
With the new priority based positioning system, when using the keynavigation
to open menus on the panel we need to follow the priority order.
The old implementation otherwise, causes the indicators menus to be shown
following the order of the ohe indicator the objects.

Bug fixed by reoreding the indicator entries considering their priority and
using the temporary generated list to select which menu should be opened.