Use a string based hash-table for geometry and don't remove values on entry-removed
When switching the focus from a window to one of its dialog children
(for example an open dialog, or any other dialog with no proper menu)
unity-panel-service gets quickly an "entry-removed" signal for an entry
and then a new "entry-added" signal for the very same entry.
This causes the panel-service not to sync with the client, since the
two events are too near and so if in this case we would have removed
the cached geometries for that entry, then the client wouldn't have shown
its menu correctly.

So now we remove the entries from the entry2geometry_hash hashtable only
when the client asked to do that to avoid errors, however we use more
memory and a slower hashtable since we need to use strings.