PanelService: improved clicks handling when indicator menus are opened
Now unity-panel-service ignores any button press event (since we can't)
check its type, and then we look at the type of the button release event.
So, when scrolling or middle-clicking over an indicator entry (that is
not an appmenu item) we send the proper signal to the indicator, and we
avoid to close the indicator menus.

This allows, for example, to change the volume using the mouse wheel when
the indicator menu is showing.

This fixes LP: #890971