BamfLauncherIcon: add a timeout on application close before removing the icon
This makes the application to correctly hide when closed, but avoids
that the launcher icon is removed when an application is quickly opened/closed

Also we override the Remove() method to make sure that once it's called
we remove the "unity-seen" flag from the BamfApplication object.
So an application icon could be re-added in the period between the Remove()
call and the ~BamfLauncherIcon call (that could take some time).