Restore the panel shadow, which has been missing since Unity 6.4. (LP: #1045705)
Unfortunately (or fortunately IMHO) this only draws the panel shadow on the
desktop. So far I have only been able to make the shadow draw on all windows
(including active) or just the desktop. This may be a restriction we have to
accept until/unless panel shadow rendering is moved into compiz itself, or
some nice person can find a way to make it work... ?