Fix singular IBus test failures due to broken or faulty connections.
First of all, we start up a new query connection for every test - this way we make sure that we have an up-to-date connection. But that's not all. We also check if the connection is correct and re-try if there are any problems, also restarting the global ibus connection when needed. This might be a bit less resource efficient, but it's not actually noticable in the global scale.