InputMonitor: add an unity class that monitors XInput2 events and converts them to XEvenet's
Clients of this class can register event handlers, and when an interested event will hit our
event filter function (that is set only if we have handlers), then we notify them with a standard
XEvent struct, converted from the XIDeviceEvent cookie.

The nice thing of this monitor is that it always reports events, despite the X grabs.