DecoratedWindow: use DecorationElement and make Edge present also in borderless windows
Windows with no border decorations might still have decoration edges so that they
can be resized properly. This was not possible before, because we were only adding
them to windows with full decorations with a top layout. Now windows can also
have only grab handles, but no borders.

Reduced also a lot the calls to verify what a window support, keeping track of
the decoration elements.