DashView: reduce delay on key-activation of first result
If user hits enter, then the first result has to be activated quickly.
So, we now wait 300ms since the live search has ended (this doesn't mean
we already have all the results, but we might have some of them) and
at this point if the user has already hit Enter key we activate the first result.
In case the user hits Enter after 300ms that the live search has ended,
we simply and promptly activate the first result.