DecoratedWindow: Just send estimated extents on _NET_REQUEST_FRAME_EXTENTS
As per EWMH protocol when _NET_REQUEST_FRAME_EXTENTS we just have to send the
client the estimation of the border. We don't have to set it on window yet.
Let's just implement this by avoiding setting the compiz borders, as this will prevent the window
to be moved to be fully visible once mapped.