UnitySettings: listen to compiz profile status in order to set low_gfx value
Instead of changing this value back and forth multiple times, just wait
this to be really changed in compiz, and update our internal variable
accordingly.
Also take care of the gsettings unity's lowgfx value only if the user has
set it.

This fixes an infinite loop when starting compiz in a lowgfx environment.